Trying to get property of non-object in filter_image_downsize
After upgrading wordpress to 4.7 and Timber to 1.1.12 I've experienced
a 'Trying to get property of non-object' error message in the
filter_image_downsize function as an attachment for the passed
attachment id does not exist.

This patch checks if the attachment returned by get_post is no
non-object and bails out of this function accordingly.